Make intro step generic
