+++ client/landing/stepper/declarative-flow/internals/steps-repository/index.ts
@@ -31,7 +31,7 @@ export { default as wooVerifyEmail } from './woo-verify-email';
 export { default as editEmail } from './edit-email';
 export { default as difmStartingPoint } from './difm-starting-point';
 export { default as letsGetStarted } from './lets-get-started';
-export { default as linkInBioIntro } from './link-in-bio-intro';
+export { default as intro } from './intro';
 export { default as chooseADomain } from './choose-a-domain';
 
 export type StepPath =
@@ -69,4 +69,4 @@ export type StepPath =
 	| 'difmStartingPoint'
 	| 'letsGetStarted'
 	| 'chooseADomain'
-	| 'linkInBioIntro';
+	| 'intro';
