+++ client/landing/stepper/declarative-flow/internals/steps-repository/intro/styles.scss
@@ -5,9 +5,16 @@ $design-button-primary-color: var( --color-primary );
 $get-started-button-primary-color: var( --studio-blue-50 );
 $white: var( --color-surface );
 
-.link-in-bio-intro {
+.intro {
 	background: url( 'calypso/assets/images/onboarding/link-in-bio-intro-background.svg' ),
 		url( 'calypso/assets/images/onboarding/link-in-bio-intro-background-noise.svg' );
+
+	&.is-newsletters {
+		background: url( 'calypso/assets/images/onboarding/paper-texture.jpg' ),
+			linear-gradient( 123deg, rgba( 7, 116, 195, 1 ) 0%, rgba( 255, 255, 255, 0 ) 100% ),
+			url( 'calypso/assets/images/onboarding/link-in-bio-intro-background-noise.svg' );
+		background-size: cover;
+	}
 	background-blend-mode: color-burn;
 	min-height: 100%;
 	min-width: 100%;
@@ -32,7 +39,7 @@ $white: var( --color-surface );
 	}
 
 	.step-container {
-		&.linkInBioIntro {
+		&.intro {
 			margin: 0;
 
 			.step-container__navigation {
@@ -41,7 +48,7 @@ $white: var( --color-surface );
 		}
 	}
 
-	.link-in-bio-intro__content {
+	.intro__content {
 		width: 100%;
 		text-align: center;
 		position: absolute;
@@ -49,7 +56,7 @@ $white: var( --color-surface );
 		transform: translateY( -50% );
 	}
 
-	.link-in-bio-intro__title {
+	.intro__title {
 		@include onboarding-font-recoleta;
 		color: $white;
 		font-weight: 500;
@@ -61,7 +68,7 @@ $white: var( --color-surface );
 		}
 	}
 
-	.link-in-bio-intro__button {
+	.intro__button {
 		color: $get-started-button-primary-color;
 		background-color: $white;
 		border: none;
